+++ src/app/types/index.ts
@@ -1,4 +1,3 @@
 export type Posts = {
-    post: string[],
-    author: string; 
+    post: string[];
 }
